JUnit 4 문서 잔재를 정리함
JUnit 4 지원 제거 이후에도 남아 있던 quickstart와 framework integration 문서를 내리고, 관련 코드 포함 예제를 JUnit Jupiter 기준으로 옮겼습니다.

데이터베이스 문서의 @Rule/@ClassRule 예시도 @Container 기반 설명으로 갱신했습니다.
