--- docs/examples/junit4/generic/build.gradle
@@ -1,12 +1,0 @@
-description = "Examples for docs"
-
-dependencies {
-    testImplementation "junit:junit:4.13.2"
-    testImplementation project(":testcontainers")
-    testImplementation project(":testcontainers-selenium")
-    testImplementation project(":testcontainers-mysql")
-
-    testRuntimeOnly 'com.mysql:mysql-connector-j:8.2.0'
-    testImplementation "org.seleniumhq.selenium:selenium-api:4.35.0"
-    testImplementation 'org.assertj:assertj-core:3.27.4'
-}
