--- docs/examples/junit4/redis/build.gradle
@@ -1,9 +1,0 @@
-description = "Examples for docs"
-
-dependencies {
-    api "io.lettuce:lettuce-core:6.8.0.RELEASE"
-
-    testImplementation "junit:junit:4.13.2"
-    testImplementation project(":testcontainers")
-    testImplementation 'org.assertj:assertj-core:3.27.4'
-}
