--- docs/examples/junit4/redis/src/test/java/quickstart/RedisBackedCacheIntTestStep0.java
@@ -1,27 +1,0 @@
-package quickstart;
-
-import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-@Ignore("This test class is deliberately invalid, as it relies on a non-existent local Redis")
-public class RedisBackedCacheIntTestStep0 {
-
-    private RedisBackedCache underTest;
-
-    @Before
-    public void setUp() {
-        // Assume that we have Redis running locally?
-        underTest = new RedisBackedCache("localhost", 6379);
-    }
-
-    @Test
-    public void testSimplePutAndGet() {
-        underTest.put("test", "example");
-
-        String retrieved = underTest.get("test");
-        assertThat(retrieved).isEqualTo("example");
-    }
-}
