+++ docs/examples/junit5/generic/src/test/java/generic/CommandsTest.java
@@ -1,15 +1,17 @@
 package generic;
 
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@Testcontainers
 public class CommandsTest {
 
-    @Rule
+    @Container
     // startupCommand {
     public GenericContainer redisWithCustomPort = new GenericContainer(DockerImageName.parse("redis:6-alpine"))
         .withCommand("redis-server --port 7777")
