+++ docs/examples/junit5/generic/src/test/java/generic/DependsOnTest.java
@@ -1,18 +1,20 @@
 package generic;
 
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@Testcontainers
 public class DependsOnTest {
 
-    @Rule
+    @Container
     // dependsOn {
     public GenericContainer<?> redis = new GenericContainer<>("redis:6-alpine").withExposedPorts(6379);
 
-    @Rule
+    @Container
     public GenericContainer<?> nginx = new GenericContainer<>("nginx:1.27.0-alpine3.19-slim")
         .dependsOn(redis)
         .withExposedPorts(80);
