+++ docs/examples/junit5/generic/src/test/java/generic/ExampleImageNameSubstitutor.java
