+++ docs/examples/junit5/generic/src/test/java/generic/ExecTest.java
@@ -1,18 +1,19 @@
 package generic;
 
-import org.junit.Rule;
-import org.junit.Test;
-import org.testcontainers.containers.Container;
+import org.junit.jupiter.api.Test;
 import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
 import java.io.IOException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@Testcontainers
 public class ExecTest {
 
-    @Rule
+    @Container
     public GenericContainer<?> container = new GenericContainer<>(DockerImageName.parse("alpine:3.17"))
         .withCommand("top");
 
@@ -23,7 +24,7 @@ public class ExecTest {
         // }
 
         // execReadingStdout {
-        Container.ExecResult lsResult = container.execInContainer("ls", "-al", "/");
+        org.testcontainers.containers.Container.ExecResult lsResult = container.execInContainer("ls", "-al", "/");
         String stdout = lsResult.getStdout();
         int exitCode = lsResult.getExitCode();
         assertThat(stdout).contains("somefile.txt");
