+++ docs/examples/junit5/generic/src/test/java/generic/support/TestSpecificImageNameSubstitutor.java
