+++ docs/examples/junit5/generic/src/test/resources/testcontainers.properties
