+++ docs/index.md
@@ -27,10 +27,9 @@ Testcontainers make the following kinds of tests easier:
 
 * Docker - please see [General Docker requirements](supported_docker_environment/index.md)
 * A supported JVM testing framework:
-    * [JUnit 4](test_framework_integration/junit_4.md) - See the [JUnit 4 Quickstart Guide](quickstart/junit_4_quickstart.md)
     * [Jupiter/JUnit 5](test_framework_integration/junit_5.md)
     * [Spock](test_framework_integration/spock.md)
-    * *Or* manually add code to control the container/test lifecycle (See [hints for this approach](test_framework_integration/junit_4.md#manually-controlling-container-lifecycle))
+    * *Or* manually add code to control the container/test lifecycle (See [manual lifecycle control](test_framework_integration/manual_lifecycle_control.md))
 
 ## Maven dependencies
